+++ src/misc/motifs.c
@@ -522,7 +522,10 @@ igraph_error_t igraph_motifs_randesu_callback(const igraph_t *graph, igraph_inte
  * the search tree.
  *
  * \param graph The graph object to study.
- * \param est Pointer to an integer, the result will be stored here.
+ * \param est Pointer to an \c igraph_real_t, the result will be stored here.
+ *        Note that even though the result is an integer, we need to use
+ *        \c igraph_real_t to avoid overflow when igraph is compiled with
+ *        32-bit integers.
  * \param size The size of the subgraphs to look for.
  * \param cut_prob Vector of probabilities for cutting the search tree
  *        at a given level. The first element is the first level, etc.
@@ -544,7 +547,7 @@ igraph_error_t igraph_motifs_randesu_callback(const igraph_t *graph, igraph_inte
  * Time complexity: TODO.
  */
 
-igraph_error_t igraph_motifs_randesu_estimate(const igraph_t *graph, igraph_integer_t *est,
+igraph_error_t igraph_motifs_randesu_estimate(const igraph_t *graph, igraph_real_t *est,
                                    igraph_integer_t size, const igraph_vector_t *cut_prob,
                                    igraph_integer_t sample_size,
                                    const igraph_vector_int_t *parsample) {
@@ -748,8 +751,10 @@ igraph_error_t igraph_motifs_randesu_estimate(const igraph_t *graph, igraph_inte
  * classes to them. Arbitrarily large motif sizes are supported.
  *
  * \param graph The graph object to study.
- * \param no Pointer to an integer type, the result will be stored
- *        here.
+ * \param no Pointer to an \c igraph_real_t, the result will be stored here.
+ *        Note that even though the result is an integer, we need to use
+ *        \c igraph_real_t to avoid overflow when igraph is compiled with
+ *        32-bit integers.
  * \param size The size of the motifs to count.
  * \param cut_prob Vector of probabilities for cutting the search tree
  *        at a given level. The first element is the first level, etc.
@@ -763,8 +768,10 @@ igraph_error_t igraph_motifs_randesu_estimate(const igraph_t *graph, igraph_inte
  * Time complexity: TODO.
  */
 
-igraph_error_t igraph_motifs_randesu_no(const igraph_t *graph, igraph_integer_t *no,
-                             igraph_integer_t size, const igraph_vector_t *cut_prob) {
+igraph_error_t igraph_motifs_randesu_no(
+    const igraph_t *graph, igraph_real_t *no, igraph_integer_t size,
+    const igraph_vector_t *cut_prob
+) {
 
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
     igraph_vector_int_t neis;